ml_project/requirements.txt
dataclasses==0.6
pyyaml==5.3.1
click==7.1.2
click-plugins==1.1.1
joblib==1.0.1 
jsonschema==3.2.0
marshmallow==3.11.1
marshmallow-dataclass==8.4.1
numpy==1.18.5
pandas==1.2.3
pytest==6.2.3
scikit-learn==0.24.1
matplotlib==3.2.1